+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+- `MigratesDatabases` trait now supports default database migrations directly in `database/migrations`.
 
 ## [0.2.7] - 2020-05-05